+++ app/Http/Controllers/Document/DocumentsController.php
@@ -26,6 +26,7 @@ use KlinkDMS\Events\UploadCompleted;
 use Klink\DmsAdapter\KlinkDocumentUtils;
 use Klink\DmsAdapter\KlinkVisibilityType;
 use Klink\DmsAdapter\Exceptions\KlinkException;
+use KlinkDMS\Jobs\ReindexDocument;
 
 class DocumentsController extends Controller
 {
@@ -840,22 +841,24 @@ class DocumentsController extends Controller
                     $document->save();
 
                     if ($document->isFileUploadComplete()) {
-                        $this->service->reindexDocument($document, KlinkVisibilityType::KLINK_PRIVATE);
+                        dispatch(new ReindexDocument($document->fresh(), KlinkVisibilityType::KLINK_PRIVATE));
                         
                         if ($user->can_capability(Capability::CHANGE_DOCUMENT_VISIBILITY)) {
                             if (! $was_document_public && $document->is_public) {
                                 \Log::info('Applying visibility change', ['descriptor' => $document->id, 'old' => $was_document_public, 'new' => $document->is_public]);
-                                $this->service->reindexDocument($document, KlinkVisibilityType::KLINK_PUBLIC);
+                                dispatch(new ReindexDocument($document->fresh(), KlinkVisibilityType::KLINK_PUBLIC));
                             } elseif ($was_document_public && ! $document->is_public) {
                                 \Log::info('Applying visibility change', ['descriptor' => $document->id, 'old' => $was_document_public, 'new' => $document->is_public]);
                                 $this->service->deletePublicDocument($document);
                             } elseif ($was_document_public && $document->is_public) {
                                 \Log::info('Reindexing also Public Document because Doc is dirty', ['descriptor' => $document->id]);
-                                $this->service->reindexDocument($document, KlinkVisibilityType::KLINK_PUBLIC);
+                                // $this->service->reindexDocument($document, KlinkVisibilityType::KLINK_PUBLIC);
+                                dispatch(new ReindexDocument($document->fresh(), KlinkVisibilityType::KLINK_PUBLIC));
                             }
                         } elseif ($was_document_public && $document->is_public) {
                             \Log::info('Reindexing also Public Document because Doc is dirty', ['descriptor' => $document->id]);
-                            $this->service->reindexDocument($document, KlinkVisibilityType::KLINK_PUBLIC);
+                            // $this->service->reindexDocument($document, KlinkVisibilityType::KLINK_PUBLIC);
+                            dispatch(new ReindexDocument($document->fresh(), KlinkVisibilityType::KLINK_PRIVATE));
                         }
                     }
                 } else {